chore: move inductee-join listener to new arch
